requirements.txt
django==4.2.16
djangorestframework==3.15.2
drf-yasg==1.21.7
channels==4.1.0
channels-redis==4.2.0
psycopg2-binary==2.9.9
gunicorn==23.0.0
redis==5.0.8